+++ core/mybatis-generator-systests-mybatis3/src/test/java/mbg/test/mb3/annotated/hierarchical/immutable/HierarchicalJava5Test.java
@@ -188,7 +188,7 @@ public class HierarchicalJava5Test extends AbstractAnnotatedHierarchicalImmutabl
 
             FieldsonlyExample example = new FieldsonlyExample();
             example.createCriteria().andIntegerfieldGreaterThan(5);
-            int rows = mapper.countByExample(example);
+            long rows = mapper.countByExample(example);
             assertEquals(2, rows);
 
             example.clear();
@@ -340,7 +340,7 @@ public class HierarchicalJava5Test extends AbstractAnnotatedHierarchicalImmutabl
 
             PkonlyExample example = new PkonlyExample();
             example.createCriteria().andIdGreaterThan(4);
-            int rows = mapper.countByExample(example);
+            long rows = mapper.countByExample(example);
             assertEquals(2, rows);
 
             example.clear();
@@ -1079,7 +1079,7 @@ public class HierarchicalJava5Test extends AbstractAnnotatedHierarchicalImmutabl
 
             PkfieldsExample example = new PkfieldsExample();
             example.createCriteria().andLastnameLike("J%");
-            int rows = mapper.countByExample(example);
+            long rows = mapper.countByExample(example);
             assertEquals(1, rows);
 
             example.clear();
@@ -1357,7 +1357,7 @@ public class HierarchicalJava5Test extends AbstractAnnotatedHierarchicalImmutabl
 
             PkblobsExample example = new PkblobsExample();
             example.createCriteria().andIdLessThan(4);
-            int rows = mapper.countByExample(example);
+            long rows = mapper.countByExample(example);
             assertEquals(1, rows);
 
             example.clear();
@@ -1656,7 +1656,7 @@ public class HierarchicalJava5Test extends AbstractAnnotatedHierarchicalImmutabl
 
             PkfieldsblobsExample example = new PkfieldsblobsExample();
             example.createCriteria().andId1NotEqualTo(3);
-            int rows = mapper.countByExample(example);
+            long rows = mapper.countByExample(example);
             assertEquals(1, rows);
 
             example.clear();
@@ -1811,7 +1811,7 @@ public class HierarchicalJava5Test extends AbstractAnnotatedHierarchicalImmutabl
 
             FieldsblobsExample example = new FieldsblobsExample();
             example.createCriteria().andFirstnameLike("S%");
-            int rows = mapper.countByExample(example);
+            long rows = mapper.countByExample(example);
             assertEquals(1, rows);
 
             example.clear();
